Renamed tableMeasures to measures
